+++ src/bbochat/forms/frm_partner_edit.py
@@ -4,6 +4,7 @@ import tkinter as tk
 from pathlib import Path
 from tkinter import ttk
 
+from bbochat.message_store import message_store
 from psiutils.constants import PAD, PADB, PADR, Mode, Status
 from psiutils.utilities import window_resize
 from psiutils.widgets import clickable_widget
@@ -12,7 +13,6 @@ from bbochat.buttons import ButtonFrame
 from bbochat.config import config
 from bbochat.constants import ICON_FILE
 from bbochat.data_store import Partner, data_store
-from bbochat.message import message_store
 
 FRAME_TITLE = "New partner"