+++ src/bbochat/forms/frm_partners.py
@@ -3,6 +3,7 @@
 import tkinter as tk
 from tkinter import ttk
 
+from bbochat.message_store import message_store
 from psiutils import messagebox
 from psiutils.constants import PAD, PADT, Mode, Status
 from psiutils.menus import Menu, MenuItem
@@ -12,7 +13,6 @@ from bbochat.buttons import ButtonFrame
 from bbochat.config import config
 from bbochat.data_store import Partner, data_store
 from bbochat.forms.frm_partner_edit import PartnerEditFrame
-from bbochat.message import message_store
 from bbochat.text import Text
 
 txt = Text()